+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 ## [v0.2.0]
 
 ### Added
+
 - `@the-devoyage/graphql-paywall` support with `paywall-helpers` apollo server plugin.
 
 ### Changed
+
 - Federation 2 Support
 - Graphql 16
-
+- Media proxy removes media serving route on proxy.